Updated the script creating approximate throughputs to fix a bug and slightly changed the function.
The script was designed to only keep parts of the baseline throughuts
that have bandpass values above a certain threshold in a continuous way.
This means that it has the throughputs in between the min and max
wavelengths where the bandpass value is above the threshold.

- In the previous version of the script, we had not included these two
min_wav and max_wav values which by definition have bandpass values
above the threshold. This has been fixed.

- Added an optional variable to enable not writing out the files.

- Some trivial changes to the function, and a clearer description of the
function